# refactor(query): optimize show drop tables