--- a/src/query/catalog/src/catalog/interface.rs
+++ b/src/query/catalog/src/catalog/interface.rs
@@ -254,12 +254,12 @@ pub trait Catalog: DynClone + Send + Sync + Debug {
     ) -> Result<Arc<dyn Table>>;
 
     // Get one table identified as dropped by db and table name.
-    async fn get_single_table_history(
+    async fn get_table_history(
         &self,
         tenant: &Tenant,
         db_name: &str,
         table_name: &str,
-    ) -> Result<Arc<dyn Table>>;
+    ) -> Result<Vec<Arc<dyn Table>>>;
 
     /// List all tables in a database.This will not list temporary tables.
     async fn list_tables(&self, tenant: &Tenant, db_name: &str) -> Result<Vec<Arc<dyn Table>>>;