I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary



1. add get_single_table_history api
2. optimize show drop tables
3. get_table_history -> get_tables_history

Performance

```
==> Running Q2: internal/queries/02.sql
select name from system.tables where name in ('t_1', 't_2');
3
Q2[1] succeeded in 0.045 seconds
Q2[2] succeeded in 0.044 seconds
Q2[3] succeeded in 0.046 seconds

==> Running Q3: internal/queries/03.sql
select name from system.tables_with_history where name in ('t_1', 't_2');
3
Q3[1] succeeded in 0.046 seconds
Q3[2] succeeded in 0.045 seconds
Q3[3] succeeded in 0.046 seconds
```

## Tests

- [x] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [x] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/16370)
<!-- Reviewable:end -->